refactor(rpc): complete partial-class split of CirclesRpcModule (#428)
* refactor(rpc): complete partial-class split of CirclesRpcModule

Finishes the in-progress partial-class refactor started in earlier commits.
Extracts the remaining methods from CirclesRpcModule.cs (3393 -> 39 LoC)
into 6 new sibling partials and appends 2 trust-summary methods to the
existing CirclesRpcModule.Trust.cs.

- New: RpcModule/CirclesRpcModule.Transactions.cs (1060 LoC) -
  GetTransactionHistory, GetTransferData, GetTransactionHistoryEnriched
  and their SQL query builders
- New: RpcModule/CirclesRpcModule.Events.cs (524 LoC) -
  GetEvents and predicate clause builders
- New: RpcModule/CirclesRpcModule.Query.cs (552 LoC) -
  Query method, query helpers, shared converters
- New: RpcModule/CirclesRpcModule.Profiles.Search.cs (438 LoC) -
  GetProfileView, SearchProfileByAddressOrName, SearchProfilesWithCursor
- New: RpcModule/CirclesRpcModule.Invitations.cs (608 LoC) -
  All 7 invitation queries and their helpers
- New: RpcModule/CirclesRpcModule.Pathfinding.cs (113 LoC) -
  FindPathV2, GetNetworkSnapshot
- Modified: RpcModule/CirclesRpcModule.Trust.cs (+123 LoC) -
  GetTrustNetworkSummary, GetAggregatedTrustRelationsEnriched appended
- Shrunk: CirclesRpcModule.cs to class shell (constants only)

Pure mechanical move. Method bodies copied byte-for-byte. Public API
surface unchanged (46 entries, byte-identical diff). Build clean
(0 errors, 61 warnings - matches pre-split baseline; the pre-existing
CS0219 unused-variable warning migrated to Query.cs:443 with its
method). Tests 147/0/113/260.

* test(rpc): widen QueryCommandTimeoutTests source grep to all partials

QueryCommandTimeoutTests is a source-text regression check that opens
CirclesRpcModule's source and greps for the Query method's NpgsqlCommand
to verify CommandTimeout is set. It hardcoded the lookup to a single
file "CirclesRpcModule.cs". With the Query method now living in
RpcModule/CirclesRpcModule.Query.cs after the partial-class split, the
grep returned -1.

Replace the single-file Path.Combine + File.ReadAllText with
Directory.GetFiles("CirclesRpcModule*.cs", SearchOption.AllDirectories)
and concatenate file contents. Test logic, assertion targets, and intent
unchanged.

Bundled with the partial-class split because that PR causes the failure.
